+++ examples/errors/main.go
@@ -13,6 +13,7 @@ func main() {
 	logger.Error().Err(mrcore.ErrInternal).Msg("this is ErrInternal")
 	logger.Error().Err(mrcore.ErrInternalTypeAssertion.New("MY-TYPE", "MY-VALUE")).Send()
 	logger.Error().Err(mrcore.ErrInternal.New().WithAttr("MY-DATA-KEY", mrmsg.Data{"itemId": "id-001"})).Send()
+	logger.Error().Err(mrcore.ErrInternalWithDetails.New("this is ErrInternalWithDetails")).Send()
 
 	logger.Fatal().Int("int1", 1).Int("int2", 2).Int("int3", 3).Msg(mrcore.ErrInternal.New().Error())
 }