+++ mrcore/usecase_errors.go
@@ -10,6 +10,7 @@ var (
 		"errUseCaseOperationFailed", mrerr.ErrorKindInternal, "operation failed")
 
 	// ErrUseCaseTemporarilyUnavailable - system is temporarily unavailable.
+	// Системная ошибка, которая сообщает о сетевых проблемах, о работоспособности внешних ресурсов (БД, API, FileSystem).
 	ErrUseCaseTemporarilyUnavailable = mrerr.NewProto(
 		"errUseCaseTemporarilyUnavailable", mrerr.ErrorKindSystem, "system is temporarily unavailable")
 
@@ -18,6 +19,7 @@ var (
 		"errUseCaseRequiredDataIsEmpty", mrerr.ErrorKindInternal, "{{ .data }} is required")
 
 	// ErrUseCaseIncorrectInputData - input data is incorrect.
+	// Это вспомогательная ошибка, для неё необязательно формировать стек вызовов и отправлять событие о её создании.
 	ErrUseCaseIncorrectInputData = mrerr.NewProto(
 		"errUseCaseIncorrectInputData", mrerr.ErrorKindInternal, "{{ .key }}={{ .data }} is incorrect")